LayoutTests/fast/media/media-query-serialization.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="help" href="http://dev.w3.org/csswg/cssom/" />
<style type="text/css" media="NOT braille, tv and (orientation: landscape) AND (min-WIDTH:100px) and (max-width: 200px ), all and (color) and (color)">
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>


<script>

description(
'Test media query serialization.  <a href="https://bugs.webkit.org/show_bug.cgi?id=39220">https://bugs.webkit.org/show_bug.cgi?id=39220</a>'
);


var expected = "not braille, tv and (max-width: 200px) and (min-width: 100px) and (orientation: landscape), (color)";
shouldBe("document.styleSheets[0].media.mediaText", "expected");

</script>

<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
