third_party/blink/web_tests/platform/generic/fast/media/media-query-serialization-expected.txt

Test media query serialization. https://bugs.webkit.org/show_bug.cgi?id=39220

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[0].media.mediaText is expected
PASS successfullyParsed is true

TEST COMPLETE
